requirements.txt
nba_api>=1.2.1
polars>=0.20.2
streamlit>=1.29.0
scikit-learn>=1.3.0
plotly>=5.18.0
pyyaml>=6.0.1
pytest>=7.4.0
black>=23.12.0
flake8>=6.1.0
joblib>=1.3.2
